+++ src/test/java/org/apache/ibatis/submitted/xml_external_ref/XmlExternalRefTest.java
@@ -66,17 +66,17 @@ public class XmlExternalRefTest {
     try {
       PersonMapper personMapper = sqlSession.getMapper(PersonMapper.class);
       Person person = personMapper.select(1);
-      assertEquals(1, person.getId());
+      assertEquals((Integer)1, person.getId());
       assertEquals(2, person.getPets().size());
-      assertEquals(2, person.getPets().get(1).getId());
+      assertEquals((Integer)2, person.getPets().get(1).getId());
 
       Pet pet = personMapper.selectPet(1);
       assertEquals(Integer.valueOf(1), pet.getId());
 
       PetMapper petMapper = sqlSession.getMapper(PetMapper.class);
       Pet pet2 = petMapper.select(3);
-      assertEquals(3, pet2.getId());
-      assertEquals(2, pet2.getOwner().getId());
+      assertEquals((Integer)3, pet2.getId());
+      assertEquals((Integer)2, pet2.getOwner().getId());
     } finally {
       sqlSession.close();
     }
